created new folder in dipy stats, moved analysis.py there
