+++ bin/dipy_ba
@@ -3,7 +3,7 @@
 from __future__ import division, print_function
 
 from dipy.workflows.flow_runner import run_flow
-from dipy.workflows.analysis import BundleAnalysisPopulationFlow
+from dipy.workflows.stats import BundleAnalysisPopulationFlow
 
 if __name__ == "__main__":
     run_flow(BundleAnalysisPopulationFlow())
